chore: add linux installer
