+++ crates/but-installer/src/shell.rs
@@ -90,11 +90,29 @@ pub(crate) fn setup_path(home_dir: &Path) -> Result<()> {
     Ok(())
 }
 
+#[cfg(target_os = "macos")]
 fn detect_shell_config(zshrc: &Path, bash_profile: &Path, bashrc: &Path) -> Option<(PathBuf, ShellType)> {
     if zshrc.exists() {
         Some((zshrc.to_path_buf(), ShellType::Zsh))
     } else if bash_profile.exists() {
         Some((bash_profile.to_path_buf(), ShellType::Bash))
+    } else if bashrc.exists() {
+        Some((bashrc.to_path_buf(), ShellType::Bash))
+    } else {
+        None
+    }
+}
+
+#[cfg(target_os = "linux")]
+fn detect_shell_config(
+    zshrc: &Path,
+    // bash_profile is only sourced when executing a login shell, which is rarely how shells are
+    // invoked on Linux distros. Therefore, we ignore it when detecting shell config on Linux.
+    _bash_profile: &Path,
+    bashrc: &Path,
+) -> Option<(PathBuf, ShellType)> {
+    if zshrc.exists() {
+        Some((zshrc.to_path_buf(), ShellType::Zsh))
     } else if bashrc.exists() {
         Some((bashrc.to_path_buf(), ShellType::Bash))
     } else {
